+++ packages/target-tauri/src-tauri/src/i18n/commands.rs
@@ -1,6 +1,6 @@
 use std::collections::HashMap;
 
-use log::error;
+use log::{error, info, warn};
 use tauri::{AppHandle, State};
 use tauri_plugin_store::StoreExt;
 use tokio::fs::read_to_string;
@@ -32,22 +32,34 @@ pub(crate) async fn get_locale_data(locale: &str, app: AppHandle) -> Result<Loca
     let locales_dir = get_locales_dir(&app).await?;
     let languages: HashMap<String, Language> = get_languages(&locales_dir).await?;
 
-    let (_locale_name, locale_dir) = match languages.get(locale) {
-        Some(Language::String(name)) => (name.to_owned(), None),
-        Some(Language::Object { name, dir }) => (name.to_owned(), Some(dir.to_owned())),
-        None => return Err(Error::LocaleNotFound(locale.to_owned())),
-    };
+    let mut get_result = languages.get(locale);
+    let mut locale_key = locale;
+
+    if get_result.is_none() {
+        warn!("locale {locale} not found, but it is a dialect/variant, so trying to fallback to base language");
+        let base_locale = locale
+            .split('-')
+            .next()
+            .ok_or(Error::BaseLocaleExtraction)?;
+        info!("base_locale: {base_locale} locale:{locale}");
+        get_result = languages.get(base_locale);
+        locale_key = base_locale;
+    }
+
+    let (_locale_name, locale_dir) = get_result
+        .ok_or(Error::LocaleNotFound(locale_key.to_owned()))?
+        .to_tuple();
 
     let untranslated_data: HashMap<String, HashMap<String, String>> =
         serde_json::from_str(&read_to_string(locales_dir.join("_untranslated_en.json")).await?)?;
 
     let language_file = {
-        let file_path = locales_dir.join(format!("{locale}.json"));
+        let file_path = locales_dir.join(format!("{locale_key}.json"));
         if file_path.exists() {
             file_path
         } else {
             error!(
-                "Unable to find language file for {locale} in {file_path:?}, defaulting to english"
+                "Unable to find language file for {locale_key} in {file_path:?}, defaulting to english"
             );
             locales_dir.join("en.json")
         }
@@ -59,7 +71,7 @@ pub(crate) async fn get_locale_data(locale: &str, app: AppHandle) -> Result<Loca
     language_data.extend(untranslated_data.into_iter());
 
     Ok(LocaleData {
-        locale: locale.to_owned(),
+        locale: locale_key.to_owned(),
         dir: locale_dir.unwrap_or_default(),
         messages: language_data,
     })