+++ packages/target-tauri/src-tauri/src/lib.rs
@@ -211,7 +211,7 @@ pub fn run() {
             let menu_manager = app.state::<MenuManger>();
             let main_window_clone = main_window.clone();
             tauri::async_runtime::block_on(menu_manager.register_window(
-                &app.handle(),
+                app.handle(),
                 &main_window,
                 Box::new(move |app| create_main_menu(app, &main_window_clone)),
             ))?;